Rework the kitchen sink
