+++ docs/kitchen-sink/admonitions.rst
@@ -1,0 +1,102 @@
+..
+   Copyright (c) 2021 Pradyun Gedam
+   Licensed under Creative Commons Attribution-ShareAlike 4.0 International License
+   SPDX-License-Identifier: CC-BY-SA-4.0
+
+===========
+Admonitions
+===========
+
+Sphinx provides several different types of admonitions.
+
+``topic``
+=========
+
+.. topic:: This is a topic.
+
+   This is what admonitions are a special case of, according to the docutils
+   documentation.
+
+``admonition``
+==============
+
+.. admonition:: The one with the custom titles
+
+   It's got a certain charm to it.
+
+``attention``
+=============
+
+.. attention::
+
+   Climate change is real.
+
+``caution``
+===========
+
+.. caution::
+
+   Cliff ahead: Don't drive off it.
+
+``danger``
+==========
+
+.. danger::
+
+   Mad scientist at work!
+
+``error``
+=========
+
+.. error::
+
+   Does not compute.
+
+``hint``
+========
+
+.. hint::
+
+   Insulators insulate, until they are subject to ______ voltage.
+
+``important``
+=============
+
+.. important::
+
+   Tech is not neutral, nor is it apolitical.
+
+``note``
+========
+
+.. note::
+
+   This is a note.
+
+``seealso``
+===========
+
+.. seealso::
+
+   Other relevant information.
+
+``tip``
+=======
+
+.. tip::
+
+   25% if the service is good.
+
+``todo``
+========
+
+.. todo::
+
+   This needs the ``sphinx.ext.todo`` extension.
+
+``warning``
+===========
+
+.. warning::
+
+   Reader discretion is strongly advised.
