--- docs/kitchen-sink/autodoc.rst
@@ -1,9 +1,0 @@
-*******
-Autodoc
-*******
-
-This is a run of automodule to grab classes, functions, and members from the
-demo module.
-
-.. automodule:: furo_demo_module
-    :members:
