+++ docs/kitchen-sink/images.rst
@@ -1,0 +1,89 @@
+..
+   Copyright (c) 2021 Pradyun Gedam
+   Licensed under Creative Commons Attribution-ShareAlike 4.0 International License
+   SPDX-License-Identifier: CC-BY-SA-4.0
+
+================
+Images & Figures
+================
+
+Images
+------
+
+An image:
+
+.. image:: https://source.unsplash.com/200x200/daily?cute+puppy
+   :height: 200
+   :width: 200
+
+A clickable image:
+
+.. image:: https://source.unsplash.com/200x200/daily?cute+puppy
+   :target: https://unsplash.com/
+   :height: 200
+   :width: 200
+
+.. image:: https://source.unsplash.com/200x200/daily?cute+puppy
+   :align: right
+   :height: 200
+   :width: 200
+
+This is a lot of text to go along with a right-aligned image, that is
+helping make this content feel less linear. It is important to have such
+a body of text, since the image is meant to be "floated" to the right,
+which would interfere with the rest of the document otherwise.
+
+Lorem ipsum dolor sit amet consectetur adipisicing elit. Blanditiis
+sapiente veritatis doloribus accusantium molestiae modi recusandae
+excepturi facere, corrupti expedita sit nihil temporibus eius sequi
+animi, illo libero labore fuga.
+
+Lorem ipsum dolor sit amet consectetur adipisicing elit. Blanditiis
+sapiente veritatis doloribus accusantium molestiae modi recusandae
+excepturi facere, corrupti expedita sit nihil temporibus eius sequi
+animi, illo libero labore fuga.
+
+.. image:: https://source.unsplash.com/200x200/daily?cute+puppy
+   :align: left
+   :height: 200
+   :width: 200
+
+This is a lot of text to go along with a left-aligned image, that is
+helping make this content feel less linear. It is important to have such
+a body of text, since the image is meant to be "floated" to the right,
+which would interfere with the rest of the document otherwise.
+
+Lorem ipsum dolor sit amet consectetur adipisicing elit. Blanditiis
+sapiente veritatis doloribus accusantium molestiae modi recusandae
+excepturi facere, corrupti expedita sit nihil temporibus eius sequi
+animi, illo libero labore fuga.
+
+Lorem ipsum dolor sit amet consectetur adipisicing elit. Blanditiis
+sapiente veritatis doloribus accusantium molestiae modi recusandae
+excepturi facere, corrupti expedita sit nihil temporibus eius sequi
+animi, illo libero labore fuga.
+
+Figures
+-------
+
+.. figure:: https://source.unsplash.com/200x200/daily?cute+puppy
+   :alt: reStructuredText, the markup syntax
+
+   A figure is an image with a caption and/or a legend:
+
+   +------------+-----------------------------------------------+
+   | re         | Revised, revisited, based on 're' module.     |
+   +------------+-----------------------------------------------+
+   | Structured | Structure-enhanced text, structuredtext.      |
+   +------------+-----------------------------------------------+
+   | Text       | Well it is, isn't it?                         |
+   +------------+-----------------------------------------------+
+
+   This paragraph is also part of the legend.
+
+A figure directive with center alignment
+
+.. figure:: https://source.unsplash.com/200x200/daily?cute+puppy
+   :align: center
+
+   This caption should be centered.
