+++ docs/kitchen-sink/index.md
@@ -1,17 +1,20 @@
+<!--
+ ~ Copyright (c) 2021 Pradyun Gedam
+ ~ Licensed under Creative Commons Attribution-ShareAlike 4.0 International License
+ ~ SPDX-License-Identifier: CC-BY-SA-4.0
+ -->
+
 # Kitchen Sink
 
-This section exists as a dump of "all sorts of edge cases" in Furo's styling.
-It is built upon an excellent base provided by Sphinx-RTD-Theme, as written by the amazing folks at Read the Docs.
+This section exists as a dump of all the things that Sphinx has.
 
-The goal is to ensure that we're adding more content here that looks good, without breaking anything externally. :)
+It has exactly one goal: to be a good checklist of things to stylise within Sphinx. This also means that it is a complete showcase of _everything_ that vanilla Sphinx provides and includes all sorts of edge cases.
+
+Some of the content here builds upon [`sphinx-rtd-theme`'s demo documentation](https://sphinx-rtd-theme.readthedocs.io/en/stable/) as well as [`docutils`' reference documentation](https://docutils.sourceforge.io/docs/index.html#ref-reference-material-for-all-groups).
 
 ```{toctree}
 :titlesonly:
+:glob:
 
-demo
-lists_tables
-really-long
-structure
-autodoc
-sphinx-design
+*
 ```
