+++ docs/kitchen-sink/typography.rst
@@ -1,0 +1,66 @@
+..
+   Copyright (c) 2021 Pradyun Gedam
+   Licensed under Creative Commons Attribution-ShareAlike 4.0 International License
+   SPDX-License-Identifier: CC-BY-SA-4.0
+
+==========
+Typography
+==========
+
+This is quite important, for a website where the majority of the content is going to be prose.
+
+Notice the font family being used for the prose, as well as the font family being used for the heading. Think about the spacing between the lines, as well as the spacing between various paragraphs. Also keep the font weight in mind, and consider if/how you want antialiasing and font-smoothing to work.
+
+Multiple paragraphs are a common occurance, because you often need more than a single paragraph to describe a thing. The rest of this paragraph is gonna be the famous lorem-ipsum: Lorem ipsum **dolor** sit amet consectetur adipisicing elit. Accusamus, sunt voluptatum tenetur libero nulla esse veritatis accusantium earum commodi hic voluptatem officia culpa optio atque. Quaerat sed quibusdam ratione nam.
+
+Headings
+========
+
+This next bit will explore how the various headings look. Think about how the content separation should work, and how the various headings should interact with the main content.
+
+Heading 3
+---------
+
+Heading 4
+^^^^^^^^^
+
+Heading 5
+~~~~~~~~~
+
+Heading 6
+*********
+
+Heading 2 with content
+======================
+
+Lorem ipsum **dolor** sit amet consectetur adipisicing elit.
+
+Accusamus, sunt voluptatum tenetur libero nulla esse veritatis accusantium earum commodi hic voluptatem officia culpa optio atque. Quaerat sed quibusdam ratione nam.
+
+Heading 3 with content
+----------------------
+
+Lorem ipsum **dolor** sit amet consectetur adipisicing elit.
+
+Accusamus, sunt voluptatum tenetur libero nulla esse veritatis accusantium earum commodi hic voluptatem officia culpa optio atque. Quaerat sed quibusdam ratione nam.
+
+Heading 4 with content
+^^^^^^^^^^^^^^^^^^^^^^
+
+Lorem ipsum **dolor** sit amet consectetur adipisicing elit.
+
+Accusamus, sunt voluptatum tenetur libero nulla esse veritatis accusantium earum commodi hic voluptatem officia culpa optio atque. Quaerat sed quibusdam ratione nam.
+
+Heading 5 with content
+~~~~~~~~~~~~~~~~~~~~~~
+
+Lorem ipsum **dolor** sit amet consectetur adipisicing elit.
+
+Accusamus, sunt voluptatum tenetur libero nulla esse veritatis accusantium earum commodi hic voluptatem officia culpa optio atque. Quaerat sed quibusdam ratione nam.
+
+Heading 6 with content
+**********************
+
+Lorem ipsum **dolor** sit amet consectetur adipisicing elit.
+
+Accusamus, sunt voluptatum tenetur libero nulla esse veritatis accusantium earum commodi hic voluptatem officia culpa optio atque. Quaerat sed quibusdam ratione nam.
